complete webui for vllm
